requirements.txt
py-cord>=2.4.0  # Updated for better slash command support
aiohttp
python-dotenv
dnspython